requirements.txt
celery[redis]==5.4.0
email-validator>=2.0.0
fastapi==0.135.1
httpx==0.28.1
jinja2>=3.1
bcrypt>=4.0.1
python-dotenv==1.2.2
python-jose[cryptography]==3.4.0
redis==5.2.1
sqlalchemy==2.0.40
uvicorn==0.41.0
